requirements.txt
pycryptodome==3.19.0
setuptools==68.0.0
wheel==0.41.2
